Create singlyLinkedList.js
This is just a rough implementation of the basic features. I still have not managed to add all the features specified in the UML in my blog post. I am currently working on re-vamping some of my old post, which is why I am implementing the linked list in JavaScript. 

I will finish up the linked list implementation as soon as possible. 

in the meanwhile, the add and remove() method should work.